Create common file for network integration scenarios
