+++ test/integration/scenarios/stress/2.register_delegate.js
@@ -28,7 +28,7 @@ var confirmTransactionsOnAllNodes = require('../../utils/transactions')
 var broadcasting = process.env.BROADCASTING !== 'false';
 
 module.exports = function(configurations) {
-	describe('stress test for type 2 transactions @slow @syncing', function() {
+	describe('Stress: type 2 transactions @slow @syncing', function() {
 		this.timeout(1800000);
 		var transactions = [];
 		var accounts = [];
